Use usize in flow_list.

+++ components/layout/flow_list.rs
@@ -102,7 +102,7 @@ impl FlowList {
 
     /// O(1)
     #[inline]
-    pub fn len(&self) -> uint {
+    pub fn len(&self) -> usize {
         self.flows.len()
     }
 }
@@ -115,7 +115,7 @@ impl<'a> Iterator for FlowListIterator<'a> {
     }
 
     #[inline]
-    fn size_hint(&self) -> (uint, Option<uint>) {
+    fn size_hint(&self) -> (usize, Option<usize>) {
         self.it.size_hint()
     }
 }
@@ -128,7 +128,7 @@ impl<'a> Iterator for MutFlowListIterator<'a> {
     }
 
     #[inline]
-    fn size_hint(&self) -> (uint, Option<uint>) {
+    fn size_hint(&self) -> (usize, Option<usize>) {
         self.it.size_hint()
     }
 }
